+++ jasmineapp/src/img/logo.png
